+++ CHANGES.rst
@@ -1,3 +1,4 @@
+* Changed ``CheckboxInput`` to correctly handle ``0`` as a value.
 * Added ``BaseFormSet.prototype.hasChanged()``
 * Changed ``Select`` widget to only allow for one selected option with the same
   value